requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Deprecated==1.2.13
distro==1.7.0
Django==4.1
django-celery-results==2.4.0
django-environ==0.9.0
django-simpleui==2022.7.29
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
flower==1.2.0
humanize==4.3.0
idna==3.3
jsonschema==3.2.0
kafka-python==2.0.2
kombu==5.2.4
openpyxl==3.0.10
packaging==21.3
paramiko==2.11.0
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.4
texttable==1.6.4
toml==0.10.2
tornado==6.2
urllib3==1.26.10
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
